requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
boto3==1.34.30
pydantic==2.5.3
pytest==7.4.3
pytest-cov==4.1.0
selenium==4.15.2
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
Flask-JWT-Extended==4.5.3
email-validator==2.1.0.post1
